requirements.txt
django>=2.2.4
pytz>=2019.2
sqlparse>=0.3.0
pillow>=6.1.0
django-modeladmin-reorder>=0.3.1
django-phone-field>=1.8.0
